Improve shooting control
Despite comment states marine should be static to shoot, it still have
a few bugs out there. This PR fix those bugs and make marine to shoot if
he is static for sure. Some condition checks are also removed since they
might not be needed to actaully do the check. If users hold the action,
shooting state will be active only if the marine stop moving.