Implement simple pincer AI to make it retaliate when shot
Refer to issue #14

Pincer now has a simple AI state machine; it's either patrolling or attacking an entity. Upon being shot by the player, it starts attacking that player. When attacking an entity, it moves towards that entity along the x-axis. In addition, the pincer dons a red tint whenever it's in its attacking state.