Pincer now doesn't instantly turn while its target is in mid-jump
The aim of this exception is to give players a fair chance to jump over the advancing pincer. If the pincer were to instantly turn around when the player is in mid-jump above the pincer, that would feel unfair to the player. With this change, the pincer only turns around after the player has landed. This makes jumping over the pincer without getting caught far more manageable.

In addition, there is now a mechanism in place for losing the target when it gets out of range. Once the target is far enough away from the pincer, the pincer loses interest and goes back to patrolling.

Lastly, remove the enum PincerState in the components/pincer file; it is unused.