Add FPS counter to top-left of screen
Currently the FPS counter is static and cannot be turned off, but it could eventually be made toggleable or be replaced by something more sophisticated.

The FPS counter allows developers to test the effects of lag on the gameplay. This is useful among other things to make the game's animations and the character's movements framerate-independent.